Create CentralEurope-3ModelAgg.ipynb
Plot multi-GCM, multi-glacier-model view of glacial runoff for Alpine river basins, 2000-2100.  Compare total runoff to each river basin, interannual variance of that runoff, and the component of runoff coming directly from glacier melt (not liquid precipitation runoff or snowmelt in a previously glacierized area).  Leaving example plots in this notebook so that they are accessible to collaborators for now.