Create SeasonalCyleComparison.ipynb
Workflow to compare seasonal cycle of glacial runoff for an example basin.  This first run analyses INDUS, where we saw some surprising SPEI values.  As a result of the figures we generated here, we identified some problems in data processing for PyGEM and GloGEM -- incorrectly shifting values from hydrologic year to calendar-year indexing -- which would affect both the seasonal cycle shown here and the SPEI computed from the runoff data.  This notebook and corresponding analyses in ehultee/glacial-SPEI will be updated once those datasets are updated.